Added Code Coverage

+++ src/main/java/com/retail/model/Item.java
@@ -1,9 +1,6 @@
 package com.retail.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @Getter
 @Setter
@@ -11,7 +8,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class Item {
 	private String name;
-	private float amount;
+	private double amount;
 	private ItemType itemType;
 
 }
