+++ src/main/java/com/retail/process/discount/AmountBasedDiscountStrategy.java
@@ -1,14 +1,14 @@
 package com.retail.process.discount;
 
 public class AmountBasedDiscountStrategy implements DiscountStrategy {
-	private final float amountCutOff;
+	private final double amountCutOff;
 	
-	public AmountBasedDiscountStrategy(float amountCutOff) {
+	public AmountBasedDiscountStrategy(double amountCutOff) {
 		this.amountCutOff = amountCutOff;
 	}
 	
 	@Override
-	public float getAmount(float totalAmount, float discountedAmount) {
+	public double getAmount(double totalAmount, double discountedAmount) {
 		return totalAmount - ((int) (totalAmount / amountCutOff) * discountedAmount);
 	}
