+++ src/main/java/com/retail/process/discount/DiscountStrategy.java
@@ -9,6 +9,6 @@ public interface DiscountStrategy {
 	 * This implementation of this method will apply the discount on total amount and return the amount after deducting the discount
 	 * Calculation of discount is based on the implementation
 	 */
-	
-	float getAmount(float totalAmount, float discount);
+
+	double getAmount(double totalAmount, double discount);
 }
